# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owl Invites API",
    "version": "0.1.0",
    "description": "Gate 2 identity, event membership, invitation household, and RSVP capability contract."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "session" },
      "invitationSessionCookie": { "type": "apiKey", "in": "cookie", "name": "owl_invitation_session" }
    },
    "schemas": {
...
      "Invitation": {
        "type": "object",
        "required": ["id", "eventId", "label", "preferredDeliveryMethod", "additionalGuestAllowance", "source", "tokenVersion", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "label": { "type": "string", "maxLength": 200 },
          "contactEmail": { "type": "string", "format": "email" },
          "contactPhone": { "type": "string" },
          "preferredDeliveryMethod": { "type": "string", "enum": ["email", "sms", "none"] },
          "additionalGuestAllowance": { "type": "integer", "minimum": 0 },
          "source": { "type": "string", "enum": ["private", "open"] },
          "tokenVersion": { "type": "integer", "minimum": 1 },
          "revokedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "InvitationGuest": {
        "type": "object",
        "required": ["id", "invitationId", "name", "origin", "attendance"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "invitationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "origin": { "type": "string", "enum": ["assigned", "additional"] },
          "attendance": { "type": "string", "enum": ["pending", "attending", "maybe", "declined"] }
        }
      },
      "InvitationQuestion": {
        "type": "object",
        "required": ["id", "label", "type", "options", "required", "scope", "sortOrder"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "label": { "type": "string" },
          "type": { "type": "string", "enum": ["text", "select", "checkbox"] },
          "options": { "type": "array", "items": { "type": "string" } },
          "required": { "type": "boolean" },
          "scope": { "type": "string", "enum": ["invitation", "guest"] },
          "sortOrder": { "type": "integer" }
        }
      },
      "InvitationHousehold": {
        "type": "object",
        "required": ["invitation", "event", "response", "guests", "questions", "invitationAnswers", "guestAnswers"],
        "properties": {
          "invitation": { "$ref": "#/components/schemas/Invitation" },
          "event": { "type": "object", "additionalProperties": true },
          "response": { "type": "object", "additionalProperties": true },
          "guests": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationGuest" } },
          "questions": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationQuestion" } },
          "invitationAnswers": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "guestAnswers": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
      "CreateInvitationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "preferredDeliveryMethod", "additionalGuestAllowance", "assignedGuestNames"],
        "properties": {
          "label": { "type": "string", "maxLength": 200 },
          "contactEmail": { "type": "string", "format": "email" },
          "contactPhone": { "type": "string" },
          "preferredDeliveryMethod": { "type": "string", "enum": ["email", "sms", "none"] },
          "additionalGuestAllowance": { "type": "integer", "minimum": 0, "maximum": 50 },
          "assignedGuestNames": { "type": "array", "minItems": 1, "items": { "type": "string" } },
          "send": { "type": "boolean" }
        }
      },
      "InvitationCapabilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["capability"],
        "properties": { "capability": { "type": "string" } }
      },
      "InvitationSubmitRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "assignedGuests", "additionalGuests", "invitationAnswers", "guestAnswers"],
        "properties": {
          "version": { "type": "integer", "minimum": 1 },
          "assignedGuests": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "additionalGuests": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "invitationAnswers": { "type": "object", "additionalProperties": true },
          "guestAnswers": { "type": "object", "additionalProperties": true }
        }
      },
      "RecoveryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["eventId", "contact"],
        "properties": {
          "eventId": { "type": "string" },
          "contact": { "type": "string" }
        }
      },
...
      "OpenEnrollmentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["capability", "label", "preferredDeliveryMethod", "guestNames"],
        "properties": {
          "capability": { "type": "string" },
          "label": { "type": "string" },
          "contactEmail": { "type": "string", "format": "email" },
          "contactPhone": { "type": "string" },
          "preferredDeliveryMethod": { "type": "string", "enum": ["email", "sms", "none"] },
          "guestNames": { "type": "array", "minItems": 1, "items": { "type": "string" } }
        }
      }
    }
  },
  "paths": {
...
    "/events/{eventId}/invitations": {
      "parameters": [{ "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": {
        "operationId": "listInvitations",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Invitation households for the event", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationHousehold" } } } } } } } }
      },
      "post": {
        "operationId": "createInvitation",
        "security": [{ "sessionCookie": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateInvitationRequest" } } } },
        "responses": { "201": { "description": "Private invitation created" } }
      }
    },
    "/events/{eventId}/invitations/{invitationId}": {
      "parameters": [
        { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
        { "name": "invitationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "operationId": "getInvitation",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Invitation household", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/InvitationHousehold" } } } } } } }
      }
    },
...
    "/invitations/exchange": {
      "post": {
        "operationId": "exchangeInvitationCapability",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationCapabilityRequest" } } } },
        "responses": { "200": { "description": "Invitation session created", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/InvitationHousehold" } } } } } } }
      }
    },
    "/invitations/session": {
      "get": {
        "operationId": "getInvitationSession",
        "security": [{ "invitationSessionCookie": [] }],
        "responses": { "200": { "description": "Current invitation household", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/InvitationHousehold" } } } } } } }
      }
    },
    "/invitations/session/response": {
      "put": {
        "operationId": "submitInvitationResponse",
        "security": [{ "invitationSessionCookie": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationSubmitRequest" } } } },
        "responses": { "200": { "description": "Response saved" }, "409": { "description": "Optimistic version conflict" } }
      }
    },
    "/invitations/recovery/request": {
      "post": {
        "operationId": "requestInvitationRecovery",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecoveryRequest" } } } },
        "responses": { "200": { "description": "Generic response regardless of match" } }
      }
    },
    "/invitations/recovery/exchange": {
      "post": {
        "operationId": "exchangeInvitationRecovery",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationCapabilityRequest" } } } },
        "responses": { "200": { "description": "Single-use recovery exchanged for invitation session" } }
      }
    },
    "/invitations/open/inspect": {
      "post": {
        "operationId": "inspectOpenEnrollment",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvitationCapabilityRequest" } } } },
        "responses": { "200": { "description": "Public event enrollment constraints" }, "404": { "description": "Unavailable without disclosure" } }
      }
    },
    "/invitations/open/enroll": {
      "post": {
        "operationId": "enrollOpenInvitation",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpenEnrollmentRequest" } } } },
        "responses": { "201": { "description": "New isolated invitation and session created" }, "409": { "description": "Capacity reached" } }
      }
    }
  }
}